update theme for Alert component, move theme to customThemeProvider
